only allow one registration per patient per clinic
If patient was already regitered, backend will send a 409.
In that case, update the patient demographics and medical
history data, but do not create new routing slip entries
or change the consent from the prior registration.

Up till now, multiple registrations created a new reg record
so we would get duplicates in the database.

Also rearranged order of updates so that routing slip and
routing slip entries come after registration is created.